fix: remove hierarchyProvider
